feat: Refactored types
