+++ src/client/organizations.ts
@@ -6,7 +6,7 @@ import {
   useQuery,
   useQueryClient
 } from '@tanstack/react-query'
-import { PaginationDto } from '@/types'
+import { PaginationDto } from '@/types/pagination-dto'
 import { OrganizationDto } from '@/types/organization-dto'
 import { deleteFetcher, getFetcher, patchFetcher, postFetcher } from './fetcher'
