+++ src/client/permissions.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import { getFetcher } from './fetcher'
-import { AuthPermissionDto } from '@/types'
+import { AuthPermissionDto } from '@/types/auth-dto'
 
 export const useGetPermissions = ({ ...options }) => {
   return useQuery<AuthPermissionDto>({
