+++ src/components/user-dropdown/index.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { useContext } from 'react'
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -11,8 +10,6 @@ import {
   DropdownMenuTrigger
 } from '../ui/dropdown-menu'
 import { Book, CircleUser, LogOut } from 'lucide-react'
-import { HeaderContext } from '../../providers/header-provider'
-import { HeaderContextType } from '../../types/header'
 import { useIntl } from '@/lib/intl/use-intl'
 import { useRouter } from 'next/navigation'
 import { useSession } from 'next-auth/react'
@@ -21,9 +18,8 @@ export const UserDropdown = () => {
   const intl = useIntl()
   const router = useRouter()
   const { data: session } = useSession()
-  const headerContext = useContext(HeaderContext) as HeaderContextType | null
 
-  const isAuthEnabled = headerContext?.isAuthEnabled !== false
+  const isAuthEnabled = process.env.PLUGIN_AUTH_ENABLED === 'true'
 
   const handleDocsClick = () => {
     window.open('https://docs.lerian.studio/', '_blank', 'noopener noreferrer')
