+++ src/types/auth-dto.ts
@@ -1,0 +1,4 @@
+type AuthResourceDto = string
+type AuthActionDto = string
+
+export type AuthPermissionDto = Record<AuthResourceDto, AuthActionDto[]>
