+++ src/types/midaz-info-dto.ts
@@ -1,0 +1,10 @@
+export enum VersionStatus {
+  UpToDate = 'up-to-date',
+  Outdated = 'outdated'
+}
+
+export type MidazInfoDto = {
+  currentVersion: string
+  latestVersion: string
+  versionStatus: VersionStatus
+}
