+++ src/types/pagination-dto.ts
@@ -1,0 +1,6 @@
+export type PaginationDto<T> = {
+  items: T[]
+  limit: number
+  page: number
+  total?: number
+}
